backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pandas==2.2.3
scikit-learn==1.5.2
pyjwt==2.8.0
bcrypt==4.1.2
psycopg2-binary==2.9.9


